[fix]: rename all classes for the corectly design pattern
