+++ src/model/DeviceModel.js
