--- src/model/status_model.js
@@ -1,9 +1,0 @@
-  class Status {
-    constructor({ id, organizationId, name, }) {
-      this.id = id;
-      this.organizationId = organizationId;
-      this.name = name;
-    }
-  }
-
-  module.exports = Status;
