+++ src/routes/index.js
@@ -1,11 +1,10 @@
 const express = require('express');
-const adminRouter = require('./admin_router.js');
-const dinningTableRouter = require('./DinningTableRouter.js');
-const deviceRouter = require('./DeviceRouter.js');
-const categoryRouter = require("./category_routes.js");
-const orderingRouter = require("./ordering_routes.js");
-const orderingProductsRouter = require("./ordering_products_routes.js");
-const organizationRouter = require("./organization_router.js");
+const adminRouter = require('./adminRoutes.js');
+const deviceRouter = require('./deviceRoutes.js');
+const categoryRouter = require("./categoryRoutes.js");
+const orderingRouter = require("./orderingRoutes.js");
+const orderingProductsRouter = require("./orderingProductsRoutes.js");
+const organizationRouter = require("./organizationRoutes.js");
 
 const router = express.Router();
 
@@ -13,7 +12,6 @@ router.use("/categories", categoryRouter);
 router.use("/ordering_products",orderingProductsRouter);
 router.use("/orderings", orderingRouter);
 router.use('/user', adminRouter);
-router.use('/dinningTable', dinningTableRouter);
 router.use('/device', deviceRouter);
 router.use('/organization', organizationRouter);
 router.use('/admin', adminRouter);
